==> analysis/integration/mc_3c/Summary/01.aggregate_confusion_matrix.ipynb <==
import pathlib
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from ALLCools.integration import calculate_diagonal_score

from wmb import cemba, aibs

## Parameters

level = 'L2'
deep_level = 'L4'
category_key = 'L3'
deep_category_key = 'L4'

## Load Annot

mc_annot = cemba.get_mc_annot()
m3c_annot = cemba.get_m3c_annot()

all_mc_cluster = mc_annot[category_key].to_pandas().value_counts().index
all_m3c_cluster = m3c_annot[category_key].to_pandas().value_counts().index

## Get L1 Confusion

def extend_confusion_matrix(matrix, mc_in_level, mc_out_level, m3c_in_level,
                            m3c_out_level):
    """Extend confusion matrix from one cluster resolution to the other deeper resolution"""
    mc_in_to_out = pd.Series(mc_annot[mc_in_level],
                             index=mc_annot[mc_out_level]).to_dict() #get the L2 and L4 identity for each cell
    mc_in_to_out = pd.Series(
        {k: v
         for k, v in mc_in_to_out.items() if v in matrix.columns}) #some l2 did not go to L4
    m3c_in_to_out = pd.Series(m3c_annot[m3c_in_level],
                              index=m3c_annot[m3c_out_level]).to_dict()
    m3c_in_to_out = pd.Series(
        {k: v
         for k, v in m3c_in_to_out.items() if v in matrix.index})

    matrix = matrix.reindex(pd.Series(mc_in_to_out).values,
                            axis=1).reindex(pd.Series(m3c_in_to_out).values)
    matrix.columns = mc_in_to_out.index
    matrix.index = m3c_in_to_out.index
    return matrix

l1_confusion_matrix = pd.read_hdf(f'../L1/ALL/L2.overlap_score.hdf')

l1_confusion_matrix = extend_confusion_matrix(l1_confusion_matrix,
                                              mc_in_level='L2',
                                              mc_out_level='L4',
                                              m3c_in_level='L2',
                                              m3c_out_level='L4')

l1_confusion_matrix.to_hdf('L1_confusion_matrx.L4_clusters.hdf', key='data')

assert l1_confusion_matrix.isna().values.sum() == 0

## Get Inte L2 Confusion

def get_inte_groups(path, annot, category):
    """read cell [mc|m3c]_integration_group.csv.gz, return cluster"""
    groups = pd.read_csv(path, index_col=0).squeeze()  # cell to group
    groups.index = groups.index.map(
        annot[category].to_pandas())  # category to group
    groups = pd.Series(groups.to_dict())  # deduplicates
    return groups #each category ro integrup

confusion_matrix_dict = {}
row_groups = {}
col_groups = {}

for path in pathlib.Path(f'../{level}/ALL').glob(f'InteGroup*/{category_key}.overlap_score.hdf'):
    group = path.parent.name
    print(group)

    # integration group
    # from leiden clustering on confusion matrix
    # may be manually merged in 07.ipynb
    row_group = get_inte_groups(path.parent / 'm3c_integration_group.csv.gz',
                                annot=m3c_annot,
                                category=category_key)
    col_group = get_inte_groups(path.parent / 'mc_integration_group.csv.gz',
                                annot=mc_annot,
                                category=category_key)
    row_groups[group] = row_group
    col_groups[group] = col_group

    # confusion matrix
    df = pd.read_hdf(path)
    df.columns = df.columns.droplevel(0)
    diag_score = calculate_diagonal_score(df,
                                          col_group=col_group,
                                          row_group=row_group) #mean score
    print(f'{group} diag score: {diag_score:.2f}')
    
    # df.columns = [col[1] for col in df.columns]
    #df = df.loc[df.index.dropna(how = 'any')]
    
    df.index.name = f'm3c.{category_key}'
    df.columns.name = f'mC.{category_key}'
    # reorder row and col based on inte groups
    df = df.loc[row_group.sort_values().index,
                col_group.sort_values().index].copy()
    confusion_matrix_dict[group] = df

all_confusion = pd.concat(confusion_matrix_dict.values())

all_confusion.to_hdf('L1_confusion_matrx.L3_clusters.hdf', key='data')

def extend_confusion_matrix(matrix, mc_in_level, mc_out_level, m3c_in_level,
                            m3c_out_level):
    """Extend confusion matrix from one cluster resolution to the other deeper resolution"""
    mc_in_to_out = pd.Series(mc_annot[mc_in_level],
                             index=mc_annot[mc_out_level]).to_dict() #get the L2 and L4 identity for each cell
    mc_in_to_out = pd.Series(
        {k: v
         for k, v in mc_in_to_out.items() if v in matrix.columns}) #some l2 did not go to L4
    m3c_in_to_out = pd.Series(m3c_annot[m3c_in_level],
                              index=m3c_annot[m3c_out_level]).to_dict()
    m3c_in_to_out = pd.Series(
        {k: v
         for k, v in m3c_in_to_out.items() if v in matrix.index})

    matrix = matrix.reindex(pd.Series(mc_in_to_out).values,
                            axis=1).reindex(pd.Series(m3c_in_to_out).values)
    matrix.columns = mc_in_to_out.index
    matrix.index = m3c_in_to_out.index
    return matrix

all_confusion_l4 = extend_confusion_matrix(all_confusion,
                                           mc_in_level='L3',
                                           mc_out_level='L4',
                                           m3c_in_level='L3',
                                           m3c_out_level='L4')
all_confusion_l4.to_hdf('L2_confusion_matrx.L4_clusters.hdf', key='data')

## Check Missing Clusters

#check prepare nb, all_mc_cluster not equal all_confusion.columns.size since cell number < 150, skipped at L2
# same for m3c_cluster

assert all_mc_cluster.size == all_confusion.columns.size 
assert all_m3c_cluster.size == all_confusion.index.size

assert all_confusion.index.duplicated().sum() == 0
assert all_confusion.columns.duplicated().sum() == 0

## Get L3/4 Integration

### Get m3c L4 to Inte L2

l3_group_names = []
for l2_group, row_group in row_groups.items():
    l3_group_name = l2_group + '_' + row_group.astype(str)
    l3_group_names.append(l3_group_name)
m3c_l3_to_inte_l2 = pd.concat(l3_group_names)

m3c_l4_to_m3c_l3 = pd.DataFrame(np.array(m3c_annot['L3']),index=np.array(m3c_annot['L4'])).squeeze()
m3c_l4_to_m3c_l3 = pd.Series(m3c_l4_to_m3c_l3.to_dict())
m3c_l4_to_inte_l2 = m3c_l4_to_m3c_l3.map(m3c_l3_to_inte_l2)

### Get mC L4 to Inte L2

l3_group_names = []
for l2_group, row_group in col_groups.items():
    l3_group_name = l2_group + '_' + row_group.astype(str)
    l3_group_names.append(l3_group_name)
mc_l3_to_inte_l2 = pd.concat(l3_group_names)

mc_l4_to_mc_l3 = pd.DataFrame(np.array(mc_annot['L3']), index=np.array(mc_annot['L4'])).squeeze()
mc_l4_to_mc_l3 = pd.Series(mc_l4_to_mc_l3.to_dict())
mc_l4_to_inte_l2 = mc_l4_to_mc_l3.map(mc_l3_to_inte_l2)

## Get Inte L4 Confusion

l4_confusion_matrix_dict = {}
l4_row_groups = {}
l4_col_groups = {}

for path in pathlib.Path(f'../{deep_level}/ALL').glob(
        f'InteGroup*/{deep_category_key}.overlap_score.hdf'):
    group = path.parent.name

    # integration group
    # from leiden clustering on confusion matrix
    # may be manually merged in 07.ipynb
    row_group = get_inte_groups(path.parent / 'm3c_integration_group.csv.gz',
                                annot=m3c_annot,
                                category=deep_category_key)
    col_group = get_inte_groups(path.parent / 'mc_integration_group.csv.gz',
                                annot=mc_annot,
                                category=deep_category_key)
    l4_row_groups[group] = row_group
    l4_col_groups[group] = col_group

    # confusion matrix
    df = pd.read_hdf(path)
    diag_score = calculate_diagonal_score(df,
                                          col_group=col_group,
                                          row_group=row_group)
    print(f'{group} diag score: {diag_score:.2f}')

    df.index.name = f'm3c.{category_key}'
    df.columns.name = f'mC.{category_key}'
    # reorder row and col based on inte groups
    df = df.loc[row_group.sort_values().index,
                col_group.sort_values().index].copy()
    l4_confusion_matrix_dict[group] = df

## Create L4 Patch on L2 Confusion Matrix

l2_with_l4_patch = all_confusion_l4.copy()

for df in l4_confusion_matrix_dict.values():
    l2_with_l4_patch.loc[df.index, df.columns] = df

l2_with_l4_patch.to_hdf('L4_confusion_matrx.L4_clusters.hdf', key='data')

## Deal with Cluster Match

l4_group_names = []
for l2_group, row_group in l4_row_groups.items():
    l4_group_name = l2_group + '_' + row_group.astype(str)
    l4_group_names.append(l4_group_name)
m3c_l4_to_inte_l4 = pd.concat(l4_group_names)

l4_group_names = []
for l2_group, row_group in l4_col_groups.items():
    l4_group_name = l2_group + '_' + row_group.astype(str)
    l4_group_names.append(l4_group_name)
mc_l4_to_inte_l4 = pd.concat(l4_group_names)

## Final Cluster Map

m3c_l4_to_final_group = {}
for m3c_cluster, l2_inte_group in m3c_l4_to_inte_l2.items():
    if m3c_cluster in m3c_l4_to_inte_l4.index:
        l4_inte_group = m3c_l4_to_inte_l4[m3c_cluster]
        #assert l4_inte_group.startswith(l2_inte_group) is True
        final_group = l4_inte_group
    else:
        final_group = l2_inte_group
    m3c_l4_to_final_group[m3c_cluster] = final_group
m3c_l4_to_final_group = pd.Series(m3c_l4_to_final_group)

m3c_l4_to_final_group.to_csv('m3c_l4_to_integration_group.csv')

mc_l4_to_final_group = {}
for mc_cluster, l2_inte_group in mc_l4_to_inte_l2.items():
    if mc_cluster in mc_l4_to_inte_l4.index:
        l4_inte_group = mc_l4_to_inte_l4[mc_cluster]
        #assert l4_inte_group.startswith(l2_inte_group) is True
        final_group = l4_inte_group
    else:
        final_group = l2_inte_group
    mc_l4_to_final_group[mc_cluster] = final_group
mc_l4_to_final_group = pd.Series(mc_l4_to_final_group)

mc_l4_to_final_group.to_csv('mc_l4_to_integration_group.csv')

m3c_l4_to_final_group

mc_l4_to_final_group
